sa segf plus...

+++ src/calc.c
@@ -212,11 +212,11 @@ void	calc(t_env *env, t_scene *scene)
 			ray[1] = 0;
 			ray[2] = 0;
 
-			coef = (((double)pix_vert - (env->height/ 2)) / (env->height/ 2)) * 0.66; //varie entre -0.66 et +0.66
-			//printf("coef %f", coef);
+			coef = (((double)pix_vert - ((double)env->height / 2)) / ((double)env->height / 2)) * 0.66; //varie entre -0.66 et +0.66
+			//printf("coef %f\n", coef);
 			ray[2] += coef * norm_vert[2];
-			coef = (((double)pix_hor - (env->width/ 2)) / (env->width/ 2)) * 0.66 * env->width_per_height; //varie entre -0.66 et +0.66
-			//printf("coef %f", coef);
+			coef = (((double)pix_hor - ((double)env->width / 2)) / ((double)env->width / 2)) * 0.66 * env->width_per_height; //varie entre -0.66 et +0.66
+			//printf("coef %f\n", coef);
 			ray[1] += coef * norm_hor[1];
 			//printf("\nray %f, %f, %f\n", ray[0], ray[1], ray[2]);
 			double	hit_sphere = calc_sphere(cam, ray);
